requirements.txt
onshape-client
pyserial
requests